POLIO-149 reorganize files, convert to ts
